# ProjectWebApi/WebApi/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WebApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Envio/FindAll": {
      "get": {
        "tags": [
          "Envio"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Envio/FindEnvioById/{id}": {
      "get": {
        "tags": [
          "Envio"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Envio/BuscarEnvioPorNumeroTracking/{numeroTracking}": {
      "get": {
        "tags": [
          "Envio"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "numeroTracking",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Envio/ListadoEnviosDetallados/{clienteId}": {
      "get": {
        "tags": [
          "Envio"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clienteId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Envio/ListadoSeguimientos/{envioId}": {
      "get": {
        "tags": [
          "Envio"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "envioId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Envio/BuscarEnvioPorFechas/{estado}": {
      "post": {
        "tags": [
          "Envio"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "estado",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BuscarEnvioPorFechasDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Envio/BuscarEnvioPorComentario/{comentario}": {
      "get": {
        "tags": [
          "Envio"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "comentario",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Usuario/FindAll": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Usuario/{id}": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Usuario/IniciarSesion": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginUsuarioDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Usuario/CrearUsuario": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "funcionarioId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AltaUsuarioDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Usuario/CambiarPassword/{id}": {
      "put": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Usuario": {
      "delete": {
        "tags": [
          "Usuario"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "usuarioId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "funcionarioId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "AltaUsuarioDTO": {
      "type": "object",
      "properties": {
        "nombreUsuario": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "rol": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "BuscarEnvioPorFechasDTO": {
      "type": "object",
      "properties": {
        "fechaInicio": {
          "format": "date",
          "type": "string"
        },
        "fechaFin": {
          "format": "date",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "LoginUsuarioDTO": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  }
}